chore(wizard-refactor): extract resolvers/subject-by-name
ResolvedSubjectMatch + SubjectResolution + resolveSubjectByName.
Drops monolith ResolvedPlaybook unused-import.
Dispatcher 11/11 + lint 17/17 green.
